extract issuance test with isolated functions and make it independent of ci provider move default issuer metadata creation into OpenID4VCI
