+++ waltid-libraries/protocols/waltid-openid4vc/src/commonMain/kotlin/id/walt/oid4vc/OpenID4VCI.kt
@@ -263,4 +263,27 @@ object OpenID4VCI {
 
         return payload
     }
+
+    fun createDefaultProviderMetadata(baseUrl: String) = OpenIDProviderMetadata(
+        issuer = baseUrl,
+        authorizationEndpoint = "$baseUrl/authorize",
+        pushedAuthorizationRequestEndpoint = "$baseUrl/par",
+        tokenEndpoint = "$baseUrl/token",
+        credentialEndpoint = "$baseUrl/credential",
+        batchCredentialEndpoint = "$baseUrl/batch_credential",
+        deferredCredentialEndpoint = "$baseUrl/credential_deferred",
+        jwksUri = "$baseUrl/jwks",
+        grantTypesSupported = setOf(GrantType.authorization_code, GrantType.pre_authorized_code),
+        requestUriParameterSupported = true,
+        subjectTypesSupported = setOf(SubjectType.public),
+        authorizationServer = baseUrl,
+        credentialIssuer = baseUrl, // (EBSI) this should be just "$baseUrl"  https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0.html#section-11.2.1
+        responseTypesSupported = setOf(
+            "code",
+            "vp_token",
+            "id_token"
+        ),  // (EBSI) this is required one  https://www.rfc-editor.org/rfc/rfc8414.html#section-2
+        idTokenSigningAlgValuesSupported = setOf("ES256"), // (EBSI) https://openid.net/specs/openid-connect-self-issued-v2-1_0.html#name-self-issued-openid-provider-
+        codeChallengeMethodsSupported = listOf("S256")
+    )
 }
