+++ waltid-libraries/protocols/waltid-openid4vc/src/commonMain/kotlin/id/walt/oid4vc/providers/OpenIDProvider.kt
@@ -1,6 +1,7 @@
 package id.walt.oid4vc.providers
 
 import id.walt.crypto.keys.Key
+import id.walt.oid4vc.OpenID4VCI
 import id.walt.oid4vc.data.*
 import id.walt.oid4vc.data.ResponseType.Companion.getResponseTypeString
 import id.walt.oid4vc.data.dif.PresentationDefinition
@@ -31,28 +32,7 @@ abstract class OpenIDProvider<S : AuthorizationSession>(
     abstract val metadata: OpenIDProviderMetadata
     abstract val config: OpenIDProviderConfig
 
-    protected open fun createDefaultProviderMetadata() = OpenIDProviderMetadata(
-        issuer = baseUrl,
-        authorizationEndpoint = "$baseUrl/authorize",
-        pushedAuthorizationRequestEndpoint = "$baseUrl/par",
-        tokenEndpoint = "$baseUrl/token",
-        credentialEndpoint = "$baseUrl/credential",
-        batchCredentialEndpoint = "$baseUrl/batch_credential",
-        deferredCredentialEndpoint = "$baseUrl/credential_deferred",
-        jwksUri = "$baseUrl/jwks",
-        grantTypesSupported = setOf(GrantType.authorization_code, GrantType.pre_authorized_code),
-        requestUriParameterSupported = true,
-        subjectTypesSupported = setOf(SubjectType.public),
-        authorizationServer = baseUrl,
-        credentialIssuer = baseUrl, // (EBSI) this should be just "$baseUrl"  https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0.html#section-11.2.1
-        responseTypesSupported = setOf(
-            "code",
-            "vp_token",
-            "id_token"
-        ),  // (EBSI) this is required one  https://www.rfc-editor.org/rfc/rfc8414.html#section-2
-        idTokenSigningAlgValuesSupported = setOf("ES256"), // (EBSI) https://openid.net/specs/openid-connect-self-issued-v2-1_0.html#name-self-issued-openid-provider-
-        codeChallengeMethodsSupported = listOf("S256")
-    )
+    protected open fun createDefaultProviderMetadata() = OpenID4VCI.createDefaultProviderMetadata(baseUrl)
 
     fun getCommonProviderMetadataUrl(): String {
         return URLBuilder(baseUrl).apply {
